写博文 需要提供写博文的功能，写博文时需要关联分类和标签。 可以上传缩略图，也可以在正文中添加图片。 文章可以直接发布，也可以保存到草稿箱。

+++ blog-framework/src/main/java/com/sangeng/domain/ArticleTag.java
@@ -20,11 +20,9 @@ import com.baomidou.mybatisplus.annotation.TableName;
 @TableName("sg_article_tag")
 public class ArticleTag  {
     //文章id@TableId
-    private Long articleId;
+    private Integer articleId;
     //标签id@TableId
-    private Long tagId;
-
-
+    private Integer tagId;
 
 
 }
